Py3k: running `futurize -0 -w .` on mseed

+++ obspy/mseed/msstruct.py
@@ -2,8 +2,13 @@
 """
 Convenience class for handling MSRecord and MSFileparam.
 """
+from __future__ import division
+from __future__ import absolute_import
+from __future__ import unicode_literals
+from future.builtins import int
+from future.builtins import open
 
-from headers import clibmseed, MSRecord, MSFileParam, MS_NOERROR, HPTMODULUS
+from .headers import clibmseed, MSRecord, MSFileParam, MS_NOERROR, HPTMODULUS
 from obspy import UTCDateTime
 import ctypes as C
 import os
@@ -29,7 +34,7 @@ def _getMSFileInfo(f, real_name):
     rec_buffer = f.read(512)
     info['record_length'] = clibmseed.ms_detect(rec_buffer, 512)
     # Calculate Number of Records
-    info['number_of_records'] = long(info['filesize'] //
+    info['number_of_records'] = int(info['filesize'] //
                                      info['record_length'])
     info['excess_bytes'] = info['filesize'] % info['record_length']
     f.seek(pos)
